+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 4.32.0
+- Add `payment_account_reference` to `Transaction::ApplePayDetails`, `Transaction::GooglePayDetails`, `Transaction::CreditCardDetails`, and `CreditCardVerification`
+- Add `processing_merchant_category_code` in `Transaction`
+- Add missing `transfer` validation error codes in `Transaction`
+
 ## 4.31.0
 
 - Add Bank Account Instant Verification Gateway with JWT creation support